add support for keeping track of subpages
when performing searches on market/aucion/cellar/trasurevault, the subpage queryparameter gets dropped, but is still available in a hidden form input. when the character switcher changes characters it will now search for that hidden input and add it to the next pages querystring so as not to get sent back to the default view for that particular page.